Fixed issue with script registration in MDS sites
As we needed to register the script with MDS in sub-sites, it was
necessary to insert the site collection url manually.

As we know that all MDS URLs include _layouts/15/start.aspx, we can use
this to extract the site collection URL and prepend the script location
with it.

For none MDS enabled sites, this is ignored anyway.